requirements.txt
hydra-core
hydra-colorlog
python-dotenv
icecream
pyrootutils == 1.0.4